Added class hierarchy, stubbed user stories as TODO's
Added the classes we talked about today in class to main.js. Added todo
tags corresponding to places people need to contribute their user
stories. In general we need to add forms to index.html to get user input
and display the game state, and fill out methods to update the state at
each turn and write changes to the display forms.