integrating more of Celestial Map into game.js
tweaked FillMapWithEncounters a litte bit: made it so that the map will
be about 75% empty space.

tweaked FillMapWithPlanets to hard-code number of encounters and
planets, rather than doing the (i+5)==11 || (i+5)==12 stuff. Also, added
the planet Eniac to this function and the TypeEnum as 15, put it at 0,0,
set it to visible.

started a feature where the CelestialMap keeps a list of visible
coordinates. We don't have to do it this way, but I thought it could be
easier to render the map from just this list, rather than iterating
through a 16k coordinates and checking each to see if they're visible.
Also, the sensor could just add coordinates to this list, rather than
having to get into the internals of the CM and update
CelestialPoint.visible values.
this list is at CelestialMap.visibleList

tweaked some values in the write_map() [formerly RenderMap] function to
line it up with the local variable names in game.js. Still printing
bizarrely many tiles.