better corner case checking for gameOver, write_prompt
game now avoids displaying multiple, out of order messages when user
might hit game over for multiple reasons in one turn. Whatever 'kiils'
the player first gets displayed. Also added a  basic write_prompt method to
write a prompt to the user display. Now, this is just used for the 'Play
Again?' prompt at game over, but we could use it in the future, to ask
user if they want to land or enter orbit, etc...