changed map size back to default 128 in initDisplay constructor
since we gave the user the option to resize the map in DEV mode, we
should build the initial map with the size parameter of 128. otherwise,
if the user starts with a small map and goes bigger, they're going to
hit array out of bounds exceptions when they try to move to coordinates
which were never made.