finished space stations. Added orbiting and docking to planets tweaked initial game state
You can now play a game of chance at space stations. Implemented
features that allow you to orbit and land on planets. Not useful yet,
but we could probably implement the recipe pretty easily and ahve a
complete game. Made changes to initial state and custom initialization
so that start state of game is consistent